Move CosmosDbService to Lib.Services
