+++ src/Lib.Services/Lib.Services.csproj
@@ -29,6 +29,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.38.0" PrivateAssets="all"/>
     <PackageReference Include="Microsoft.Extensions.Logging" Version="$(CorePackageVersion)" PrivateAssets="all"/>
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="$(CorePackageVersion)" PrivateAssets="all"/>
     <PackageReference Include="Microsoft.Extensions.Http" Version="$(CorePackageVersion)" PrivateAssets="all"/>
