+++ src/Lib.Services/Logging/CosmosDb/CosmosDbServiceLogging.cs
@@ -1,6 +1,6 @@
 using Microsoft.Extensions.Logging;
 
-namespace MuzakBot.App.Logging.CosmosDb;
+namespace MuzakBot.Lib.Services.Logging.CosmosDb;
 
 /// <summary>
 /// Source generated logging methods for the <see cref="Services.CosmosDbService"/> class.
