+++ src/Lib.Services/Services/CosmosDbService/LyricsAnalyzer/AddOrUpdateLyricsAnalyzerConfigAsync.cs
@@ -1,11 +1,11 @@
 using System.Net;
 using Microsoft.Azure.Cosmos;
 using Microsoft.Extensions.Logging;
-using MuzakBot.App.Extensions;
-using MuzakBot.App.Logging.CosmosDb;
+using MuzakBot.Lib.Services.Extensions.Telemetry;
+using MuzakBot.Lib.Services.Logging.CosmosDb;
 using MuzakBot.Lib.Models.Database.LyricsAnalyzer;
 
-namespace MuzakBot.App.Services;
+namespace MuzakBot.Lib.Services;
 
 public partial class CosmosDbService
 {
